+++ Plugins/BatchProcessing/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/batchprocessing/batch/Request.java
@@ -145,7 +145,7 @@ public abstract class Request implements Content {
      *
      * @return the headers of the individual batch request.
      */
-    public Map<String, String> getHttpHeaders() {
+    public Map<String, String> getInnerHeaders() {
         return headersInner;
     }