+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Development version 2.2.1
 
+**New Features**
+* Added support for headers in JSON-Batch requets, enabling JSON-Patch requests in JSON-Batch requests.
+
 
 ## Release version 2.2.0